Fixed the worst of the memory leak
Every time a jump was executed sizeof(BasicBlock) was malloc'ed.
By moving the malloc call into the branch that creates a new BasicBlock
the worst of the memory leak should be fixed.
The other minor problem with the BasicBlocks never getting removed still
persists as long as the further logic is not implemented.