Fixed coredump at second block creation
Ran some tests and the logs showed some strange results.
188 blocks were marked for JIT'ing but the instructions per block varied
massively.
Some of the Blocks contain upwards of 10 million instructions while
the vast majority contain less than 1000.
At this point its unknown if this is due to an oversight on my part or
if the blocks are truly this massive.