Somewhat got the basicblocks working again
BasicBlock detection and heating up seems to kinda work now.
Its horribly slow since i moved from a map to a vector, that's a todo
for the future.
The "fix" was to get rid of the `Location` struct and just manage the
BasicBlocks by a string - is it clean? no, is it inefficient, probably,
do i hate it, oh absolutely, but it DOES work for a change, which is
  nice. \o/