Converted the Location pointers back to strings
As before, for some (to me) unknown reason the pointers for the
Location structs keep getting messed up.
The quick and dirty solution is to once again use strings. The whole
purpose of these Location structs is mainly debugging in the first place
(ignoring them being the index for the map of basicBlocks, but this also
works fine with strings as key instead of structs).